Day 5 answer
Yes, I hardcoded my input's stack. Blame Reddit, not me: I was coding the manual parser and I saw a meme about how it took someone 3x longer than the actual solution, and knowing how slowly I code for these, that meant infinite pain and suffering, so I made the express decision to not code it.

Could I have stolen the code from someone else? Yes, but I wanna make it clear that I did not, in fact, code this, and took the easy path of hardcoding it.